Move CLI to a separate submodule, add empty config module
Also, read DEBUG_SYSFONTS and DEBUG_UNSUPPORTED env vars in
setup_logging() and use pdfium_i._safe_debug() in sysfontinfo where we
have to.
