+++ src/pypdfium2_cfg/__init__.py
@@ -1,0 +1,4 @@
+# SPDX-FileCopyrightText: 2026 geisserml <geisserml@gmail.com>
+# SPDX-License-Identifier: Apache-2.0 OR BSD-3-Clause
+
+# TODO
