+++ src/pypdfium2_cli/extract_images.py
@@ -8,7 +8,7 @@ import traceback
 from pathlib import Path
 import pypdfium2.raw as pdfium_c
 import pypdfium2._helpers as pdfium
-from pypdfium2._cli._parsers import (
+from pypdfium2_cli._parsers import (
     add_input, get_input,
     BooleanOptionalAction,
 )
