+++ src/pypdfium2_cli/extract_text.py
@@ -1,7 +1,7 @@
 # SPDX-FileCopyrightText: 2026 geisserml <geisserml@gmail.com>
 # SPDX-License-Identifier: Apache-2.0 OR BSD-3-Clause
 
-from pypdfium2._cli._parsers import add_input, get_input
+from pypdfium2_cli._parsers import add_input, get_input
 
 EXTRACT_RANGE   = "range"
 EXTRACT_BOUNDED = "bounded"
