+++ src/pypdfium2_cli/pdfinfo.py
@@ -3,7 +3,7 @@
 
 import pypdfium2.raw as pdfium_c
 import pypdfium2.internal as pdfium_i
-from pypdfium2._cli._parsers import (
+from pypdfium2_cli._parsers import (
     add_input,
     add_n_digits,
     get_input,
