+++ src/pypdfium2_cli/toc.py
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: Apache-2.0 OR BSD-3-Clause
 
 import pypdfium2.internal as pdfium_i
-from pypdfium2._cli._parsers import (
+from pypdfium2_cli._parsers import (
     add_input,
     add_n_digits,
     get_input,
